moved entity-tag types & fns to their own files
