+++ client/src/app/components/evidence/evidence-tag/evidence-tag.component.ts
@@ -10,7 +10,7 @@ import {
 } from '@angular/core'
 
 import { getEntityColor } from '@app/core/utilities/get-entity-color'
-import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.component'
+import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.types'
 import { EvidenceStatus, Maybe } from '@app/generated/civic.apollo'
 import { NzPopoverDirective } from 'ng-zorro-antd/popover'
