+++ client/src/app/components/features/feature-tag/feature-tag.component.ts
@@ -9,7 +9,7 @@ import {
 } from '@angular/core'
 
 import { getEntityColor } from '@app/core/utilities/get-entity-color'
-import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.component'
+import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.types'
 import { Maybe } from '@app/generated/civic.apollo'
 import { NzPopoverDirective } from 'ng-zorro-antd/popover'
