+++ client/src/app/components/molecular-profiles/molecular-profile-tag/molecular-profile-tag.component.ts
@@ -8,7 +8,7 @@ import {
 } from '@angular/core'
 
 import { getEntityColor } from '@app/core/utilities/get-entity-color'
-import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.component'
+import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.types'
 import { MolecularProfileSegment } from '@app/generated/civic.apollo'
 import { NzPopoverDirective } from 'ng-zorro-antd/popover'
