+++ client/src/app/components/organizations/organization-tag/organization-tag.component.ts
@@ -5,7 +5,7 @@ import {
   QueryList,
   ViewChildren,
 } from '@angular/core'
-import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.component'
+import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.types'
 import { NzPopoverDirective } from 'ng-zorro-antd/popover'
 
 export interface TagLinkableOrganization {
