+++ client/src/app/components/revisions/revision-tag/revision-tag.component.ts
@@ -5,7 +5,7 @@ import {
   QueryList,
   AfterViewInit,
 } from '@angular/core'
-import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.component'
+import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.types'
 import { Maybe } from '@app/generated/civic.apollo'
 import { NzPopoverDirective } from 'ng-zorro-antd/popover'
