+++ client/src/app/components/variant-groups/variant-group-tag/variant-group-tag.component.ts
@@ -7,7 +7,7 @@ import {
   ViewChildren,
 } from '@angular/core'
 import { getEntityColor } from '@app/core/utilities/get-entity-color'
-import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.component'
+import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.types'
 import { Maybe } from '@app/generated/civic.apollo'
 import { NzPopoverDirective } from 'ng-zorro-antd/popover'
