+++ client/src/app/components/variants/variant-tag/variant-tag.component.ts
@@ -10,7 +10,7 @@ import {
 } from '@angular/core'
 
 import { getEntityColor } from '@app/core/utilities/get-entity-color'
-import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.component'
+import { PopoverPlacement } from '@app/forms/components/entity-tag/entity-tag.types'
 import { NzPopoverDirective } from 'ng-zorro-antd/popover'
 
 export interface LinkableVariant {
