+++ client/src/app/forms/components/entity-collection-tag/entity-collection-tag.component.ts
@@ -5,14 +5,14 @@ import {
   TemplateRef,
 } from '@angular/core'
 import { CvcEntityTagListConfig } from '../entity-tag-list/entity-tag-list.component'
-import { LinkableEntity } from '../entity-tag/entity-tag.component'
+import { LinkableEntity } from '../entity-tag/entity-tag.types'
 
 @Component({
-    selector: 'cvc-entity-collection-tag',
-    templateUrl: './entity-collection-tag.component.html',
-    styleUrls: ['./entity-collection-tag.component.less'],
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    standalone: false
+  selector: 'cvc-entity-collection-tag',
+  templateUrl: './entity-collection-tag.component.html',
+  styleUrls: ['./entity-collection-tag.component.less'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: false,
 })
 export class CvcEntityCollectionTagComponent {
   @Input() set cvcCollectionTagConfig(config: CvcEntityTagListConfig) {
