+++ client/src/app/forms/components/entity-tag/entity-tag.functions.ts
@@ -1,0 +1,10 @@
+import { TypeGuard } from '@app/core/pipes/type-guard.pipe'
+import { LinkableEntity } from './entity-tag.types'
+
+export const isLinkableEntity: TypeGuard<any, LinkableEntity> = (
+  entity: any
+): entity is LinkableEntity =>
+  entity !== undefined &&
+  entity.__typename &&
+  entity.id &&
+  entity.name !== undefined
