+++ client/src/app/forms/types/molecular-profile-select/mp-expression-editor/mp-expression-editor.component.ts
@@ -23,7 +23,7 @@ import {
   MutationState,
   MutatorWithState,
 } from '@app/core/utilities/mutation-state-wrapper'
-import { LinkableEntity } from '@app/forms/components/entity-tag/entity-tag.component'
+import { LinkableEntity } from '@app/forms/components/entity-tag/entity-tag.types'
 import {
   CreateMolecularProfile2GQL,
   CreateMolecularProfile2Mutation,
@@ -67,11 +67,11 @@ type ExampleExpression = {
 
 @UntilDestroy()
 @Component({
-    selector: 'cvc-mp-expression-editor',
-    templateUrl: './mp-expression-editor.component.html',
-    styleUrls: ['./mp-expression-editor.component.less'],
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    standalone: false
+  selector: 'cvc-mp-expression-editor',
+  templateUrl: './mp-expression-editor.component.html',
+  styleUrls: ['./mp-expression-editor.component.less'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: false,
 })
 export class MpExpressionEditorComponent implements AfterViewInit, OnChanges {
   @Input() cvcPrepopulateWithId: Maybe<number>
@@ -164,7 +164,7 @@ export class MpExpressionEditorComponent implements AfterViewInit, OnChanges {
       description: 'KIT D816V must be absent.',
     },
   ]
-  
+
   viewer$: Observable<Viewer>
   mostRecentOrg$: Observable<Maybe<Organization>>
   mostRecentOrgId: Maybe<number>
@@ -174,7 +174,7 @@ export class MpExpressionEditorComponent implements AfterViewInit, OnChanges {
     private createMolecularProfileGql: CreateMolecularProfile2GQL,
     private mpEditorPrepopulate: MpExpressionEditorPrepopulateGQL,
     private networkErrorService: NetworkErrorsService,
-    private viewerService: ViewerService,
+    private viewerService: ViewerService
   ) {
     this.createMolecularProfileMutator = new MutatorWithState(
       this.networkErrorService
@@ -198,13 +198,9 @@ export class MpExpressionEditorComponent implements AfterViewInit, OnChanges {
   }
 
   ngAfterViewInit(): void {
-    this.mostRecentOrg$
-      .pipe(
-        untilDestroyed(this)
-      )
-      .subscribe((org) => {
-        this.mostRecentOrgId = org?.id
-      })
+    this.mostRecentOrg$.pipe(untilDestroyed(this)).subscribe((org) => {
+      this.mostRecentOrgId = org?.id
+    })
     this.onInputChange$
       .pipe(
         // tag('onInputChange$'),
