+++ client/src/app/forms/types/variant-select/variant-manager/variant-manager.types.ts
@@ -1,9 +1,8 @@
-import { TemplateRef } from '@angular/core'
 import { ApolloError } from '@apollo/client/core'
 import { EnumOutputStyle } from '@app/core/pipes/evidence-enum-display-type'
 import { TypeGuard } from '@app/core/pipes/type-guard.pipe'
 import { CvcEmptyValueCategory } from '@app/forms/components/empty-value/empty-value.component'
-import { CvcTagLabelMax } from '@app/forms/components/entity-tag/entity-tag.component'
+import { CvcTagLabelMax } from '@app/forms/components/entity-tag/entity-tag.types'
 import {
   VariantManagerFieldsFragment,
   VariantManagerQuery,
@@ -16,11 +15,9 @@ import {
 } from '@app/generated/civic.apollo'
 import { GraphQLFormattedError } from 'graphql'
 import {
-  NzTableFilterFn,
   NzTableFilterList,
   NzTableFilterValue,
   NzTableQueryParams,
-  NzTableSortFn,
   NzTableSortOrder,
 } from 'ng-zorro-antd/table'
 import { Subject } from 'rxjs'
